requirements.txt
pandas
numpy
requests==2.23.0
flask
tensorflow==2.5.1
Pillow
scikit-learn
